pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Mashal", email = "Mashal.Jawed@imperiumdynamics.com"},
]
dependencies = ["python-dotenv>=1.2.2", "langchain>=1.2.13", "langchain-openai>=1.1.12", "faiss-cpu>=1.13.2", "langchain-community>=0.4.1", "tiktoken>=0.12.0", "langchain-text-splitters>=1.1.1", "openai>=2.30.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
